requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Markdown==0.3
idna==3.3
importlib-metadata==4.10.0
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
python-dotenv==0.19.2
python-podcastindex==1.7.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
